move contact, external_accession, project_summary and publication serializers to their own files
