+++ api/scpca_portal/serializers/__init__.py
@@ -1,4 +1,8 @@
 from .computed_file import ComputedFileSerializer
+from .contact import ContactSerializer
+from .external_accession import ExternalAccessionSerializer
+from .project_summary import ProjectSummarySerializer
+from .publication import PublicationSerializer
 from .serializers import (
     DatasetCreateSerializer,
     DatasetDetailSerializer,
@@ -6,6 +10,5 @@ from .serializers import (
     DatasetUpdateSerializer,
     ProjectLeafSerializer,
     ProjectSerializer,
-    ProjectSummarySerializer,
     SampleSerializer,
 )
