+++ api/scpca_portal/serializers/external_accession.py
@@ -1,0 +1,13 @@
+from rest_framework import serializers
+
+from scpca_portal.models import ExternalAccession
+
+
+class ExternalAccessionSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = ExternalAccession
+        fields = (
+            "accession",
+            "has_raw",
+            "url",
+        )
