+++ api/scpca_portal/serializers/publication.py
@@ -1,0 +1,13 @@
+from rest_framework import serializers
+
+from scpca_portal.models import Publication
+
+
+class PublicationSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Publication
+        fields = (
+            "citation",
+            "doi",
+            "doi_url",
+        )
